// kbeatz-ui/src/api/generated/models/ScanErrorEntry.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ScanErrorEntry = {
    /**
     * Album directory path relative to the library root
     */
    albumDir: string;
    /**
     * Human-readable error reason without stack traces or absolute paths
     */
    reason: string;
    /**
     * Actionable suggestion for the user
     */
    suggestion: string;
};